Added test for cancelling
Test states a student enlists in a section, it then checks if a student has one section enlisted and if the aforementioned section has one student. Student then cancels then the tests checks the amount of students in the class. Test also handles cancelling on a section a student doesn't have enlisted on, which should return false.